requirements/production.txt
dependency_injector==4.42.0
Django==5.1.1
matplotlib==3.9.2
numpy==2.1.2
python-decouple==3.8
sympy==1.13.3
scipy==1.8.0
